[UPDATE] moved FetchRecords to PersistenManager+fetch and same with test cases
