+++ Sources/YPersistence/Manager/PersistentManager+Fetch.swift
@@ -25,6 +25,36 @@ extension PersistenceManager {
         try self.fetchRecords(predicate: predicate, sortDescriptors: sortDescriptors, context: nil)
     }
 
+    /// Fetches managed object records from Core Data.
+    /// May be safely called from any thread, but the managed objects returned
+    /// may only be safely accessed from the current thread.
+    /// - Parameters:
+    ///   - predicate: optional predicate to filter records
+    ///   - sortDescriptors: optional sort to apply to fetched records
+    ///   - context: context to fetch data
+    /// - Returns: an array of managed object records matching the (optional) predicate and optionally sorted
+    /// - Throws: any error executing the fetch request
+    internal func fetchRecords<T: DataRecord>(
+        predicate: NSPredicate? = nil,
+        sortDescriptors: [NSSortDescriptor]? = nil,
+        context: NSManagedObjectContext? = nil
+    ) throws -> [T] {
+        let context = context ?? contextForThread()
+        let fetchRequest = NSFetchRequest<T>(entityName: T.entityName)
+
+        if let predicate = predicate {
+            fetchRequest.predicate = predicate
+        }
+
+        if let sortDescriptors = sortDescriptors {
+            fetchRequest.sortDescriptors = sortDescriptors
+        }
+
+        fetchRequest.returnsObjectsAsFaults = false
+
+        return try context.fetch(fetchRequest)
+    }
+
     /// Fetches records from Core Data and convert those to models.
     /// May be safely called from any thread, and the models returned may be safely transferred to any thread.
     /// - Parameters:
